chore: move unit tests under own folder
