+++ beet/package.json
@@ -12,8 +12,8 @@
     "postpublish": "git push origin && git push origin --tags",
     "build": "rimraf dist && tsc -p tsconfig.json",
     "pretest": "yarn build",
-    "test": "tape dist/test/*.js",
-    "test:browser": "airtap ./test/*.ts",
+    "test": "tape dist/test/{unit,compat}/*.js",
+    "test:browser": "airtap ./test/{unit,compat}/*.ts",
     "lint": "prettier -c ./src/",
     "lint:fix": "prettier --write ./src",
     "doc": "rimraf ../docs/beet && typedoc",
