+++ beet/test/unit/collections.array.ts
@@ -4,7 +4,7 @@ import {
   fixedSizeUtf8String,
   FixedSizeBeet,
   u8,
-} from '../src/beet'
+} from '../../src/beet'
 import test from 'tape'
 
 function checkCases<T>(
